requirements.txt
pandas
numpy
joblib
scikit-learn==1.1.1
Flask
tensorflow-intel
gunicorn
